# [SPARK-5938][SPARK-5443][SQL] Improve JsonRDD performance