--- a/sql/core/src/main/scala/org/apache/spark/sql/SQLContext.scala
+++ b/sql/core/src/main/scala/org/apache/spark/sql/SQLContext.scala
@@ -616,7 +616,7 @@ class SQLContext(@transient val sparkContext: SparkContext)
   @Experimental
   def jsonRDD(json: RDD[String], schema: StructType): DataFrame = {
     if (conf.useJacksonStreamingAPI) {
-      baseRelationToDataFrame(new JSONRelation(json, None, 1.0, Some(schema))(this))
+      baseRelationToDataFrame(new JSONRelation(() => json, None, 1.0, Some(schema))(this))
     } else {
       val columnNameOfCorruptJsonRecord = conf.columnNameOfCorruptRecord
       val appliedSchema =
@@ -650,7 +650,7 @@ class SQLContext(@transient val sparkContext: SparkContext)
   @Experimental
   def jsonRDD(json: RDD[String], samplingRatio: Double): DataFrame = {
     if (conf.useJacksonStreamingAPI) {
-      baseRelationToDataFrame(new JSONRelation(json, None, samplingRatio, None)(this))
+      baseRelationToDataFrame(new JSONRelation(() => json, None, samplingRatio, None)(this))
     } else {
       val columnNameOfCorruptJsonRecord = conf.columnNameOfCorruptRecord
       val appliedSchema =