… sql/core/src/main/scala/org/apache/spark/sql/json/JacksonGenerator.scala
… sql/core/src/main/scala/org/apache/spark/sql/json/JacksonParser.scala
… sql/core/src/main/scala/org/apache/spark/sql/json/JacksonUtils.scala
… sql/core/src/main/scala/org/apache/spark/sql/json/JsonRDD.scala
… sql/core/src/test/scala/org/apache/spark/sql/json/JsonSuite.scala
… sql/core/src/test/scala/org/apache/spark/sql/sources/InsertSuite.scala