@rxin 2015-04-30
I won't have time to look at this today, but this is pretty cool.