@NathanHowell 2015-04-30
Looks like it may also resolve [SPARK-5443](https://issues.apache.org/jira/browse/SPARK-5443).