@rxin 2015-04-30
Can you put both JIRA tickets in the title? It will then automatically linked to both tickets.