@NathanHowell 2015-04-30
Benchmarked a small-ish real dataset... Runs are with 5 executors (for 5 input splits) with data in HDFS:

step | before | after 
------|----------|--------
`val df = sqlContext.jsonRDD(...)` - schema inference | 37.14s | 18.16s | 
`df.count()` | 125.8s | 25.7s
`df.select("col1").count()` | 96.9s | 26.5s

Not sure why but the new code seems a bit slower when using projection pushdowns. It may be schema dependent or overhead from evaluating the projection expression.