@marmbrus 2015-05-01
I'm okay with freezing the old code and not having tests.  I just want a
quick fall back if a regression is found.
On May 1, 2015 4:01 PM, "Nathan Howell" <notifications@github.com> wrote:

> @yhuai <https://github.com/yhuai> The updated patches do not test the old
> code. Do you have an opinion on the best way to address this? I can
> duplicate the entire JsonSuite or try to do something a bit better...
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/apache/spark/pull/5801#issuecomment-98258602>.
>