@yhuai 2015-05-04
@NathanHowell It looks good! Just a few minor comments.