@NathanHowell 2015-05-04
I split up the rather large `JsonRDD2` object into multiple objects. I picked names that match the Jackson class names but I'm happy to rename them if you have something else in mind.

* `JacksonParser: String => Row`
* `JacksonGenerator: Row => String`
* `InferSchema: String => StructType`