@rxin 2015-05-04
This looks good to me, but I didn't go through the PR in detail. I will let @yhuai do the final check. Thanks a lot. This is going to be amazing for JSON users!