@rxin 2015-05-04
@NathanHowell In preparation for merging, do you mind updating the pull request description to include the latest perf numbers, and make sure the old description is still valid?