@NathanHowell 2015-05-04
@rxin Do you know if there are any semi-official datasets around for benchmarking? I think it would be better to publish reproducible numbers if possible.