@rxin 2015-05-04
You can try get some tweets and use that. That's what I usually demo on, but unfortunately I don't think it is legal to make collected tweets public.

There are some stuff here: https://www.opensciencedatacloud.org/publicdata/city-of-chicago-public-datasets/