@NathanHowell 2015-05-05
@yhuai Is there still time to get this in for 1.4.0?