@marmbrus 2015-05-05
Yeah, given that there is a flag I think we can still include this.