@yhuai 2015-05-05
@NathanHowell I will take a final check tomorrow. Can you also add the performance number of selecting all columns in the description? You can use `df.rdd.count` as the command to compare two versions.