@NathanHowell 2015-05-05
@yhuai I'll be able to check on this a bit later today.